+++ lib/models/response/customer.dart
@@ -13,6 +13,8 @@
 // IMPLIED WARRANTIES, INCLUDING, WITHOUT LIMITATION, THE IMPLIED
 // WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE.
 
+import '../meta_data.dart';
+
 class Customer {
   int id;
   String dateCreated;
@@ -100,21 +102,6 @@ class Customer {
     }
     return data;
   }
-}
-
-class MetaData {
-  final int id;
-  final String key;
-  final String value;
-
-  MetaData(this.id, this.key, this.value);
-
-  MetaData.fromJson(Map<String, dynamic> json)
-      : id = json['id'],
-        key = json['key'],
-        value = json['value'];
-
-  Map<String, dynamic> toJson() => {'id': id, 'key': key, 'value': value};
 }
 
 class Billing {