+++ lib/models/response/product_variation.dart
@@ -13,6 +13,8 @@
 // IMPLIED WARRANTIES, INCLUDING, WITHOUT LIMITATION, THE IMPLIED
 // WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE.
 
+import '../meta_data.dart';
+
 class ProductVariation {
   int id;
   DateTime dateCreated;
@@ -135,21 +137,6 @@ class ProductVariation {
         metaData = (json['meta_data'] as List)
             .map((i) => MetaData.fromJson(i))
             .toList();
-}
-
-class MetaData {
-  final int id;
-  final String key;
-  final String value;
-
-  MetaData(this.id, this.key, this.value);
-
-  MetaData.fromJson(Map<String, dynamic> json)
-      : id = json['id'],
-        key = json['key'],
-        value = json['value'].toString();
-
-  Map<String, dynamic> toJson() => {'id': id, 'key': key, 'value': value};
 }
 
 class Category {